Graphics.Camera

Higher-level camera functions

viewerRotation :: Camera3D c => c a -> V2 a -> V2 a -> Quaternion a

Given a line segment in viewport coordinates, compute an approximate camera rotation.

viewerMovement :: Camera3D c => c a -> V2 a -> V2 a -> V3 a

Given a line segment in viewport coordinates, compute an approximate camera translation.

viewVector :: (Epsilon a, RealFloat a, Camera3D c) => c a -> V2 a -> V3 a

Convert a point in viewport coordinates into a world-space vector pointing away from the camera.

vectorView :: (Epsilon a, RealFloat a, Camera3D c) => c a -> V3 a -> V2 a

Convert a world-space vector into viewport coordinates if it's visible. Vectors pointing backwards (behind the camera) will be treated equivalent to their negation. Vectors orthogonal to the camera view will produce infinity or NaN coordinates.

centerView :: (Epsilon a, Ord a, RealFloat a, Camera3D c) => c a -> V2 a -> c a

Given a point in viewport coordinates, rotate the camera to center on it.

centerZoomView :: (Ord a, Num a, Camera3D c) => c a -> (V2 a, V2 a) -> c a

Given a rectangle in viewport coordinates, rotate the camera to center on it and zoom in as much as possible.

centerZoomOn :: Camera3D c => c a -> (V3 a, V3 a) -> c a

Given a box in world-space coordinates, rotate the camera to center on it and zoom in as much as possible.

centerOn :: Camera3D c => c a -> V3 a -> c a

Given a point in world-space coordinates, rotate the camera to center on it.

viewerSize :: (Ord a, Epsilon a, RealFloat a, Camera3D c) => c a -> (V3 a, V3 a) -> (V2 a, V2 a)

Given a box in world-space coordinates, calculate how large it would be in viewport coordinates.

viewSpace :: Camera3D c => c a -> [[V3 a]]

Get the world-space bounds of the camera's view as a 3D mesh.

dolly :: (JibCamera c, GimbalCamera c, Epsilon a, Conjugate a, RealFloat a) => V3 a -> c a -> c a

Interpret an offset in the camera's perspective with gimbal applied, scaled proportionally to jib distance, then move the camera in world space.

If you have a jib & gimbal set up to orbit the camera around a center point, this should make the center point move "naturally" with respect to the rendered scene.

vertigo :: (JibCamera c, PerspectiveCamera c, Floating a) => a -> c a -> c a

Perform a dolly zoom, i.e. change the focal length and distance simultaneously so that whatever is at the base position of the camera remains the same size as the perspective changes.

Try it out if you're wondering why it's named "vertigo".

Camera interface classes

class Camera c where

The Camera class is a basic common interface to different cameras.

Note: The result of setting camera properties to values that do not satisfy the requirements specified for each field is undefined behavior. In particular, this may result in throwing exceptions and/or violating the lens laws.

Minimal complete definition

viewArea

Methods

viewArea :: Lens' (c a) (V2 a)

The dimensions of the image seen by the camera, in view space coordinates. Generally this will be the pixel size of the rendered image.

Both dimensions should be positive, non-zero values.

viewWidth and viewHeight are the X and Y components

viewAspect is the aspect ratio W/H

viewDiagonal is the diagonal size of the view area

viewWidth :: Lens' (c a) a

The width of the image seen by the camera, in view space coordinates.

viewWidth should be a positive, non-zero value.

Accesses the X component of viewArea

Leaves viewHeight unchanged

Adjusts viewDiagonal and viewAspect to match the new width.

viewHeight :: Lens' (c a) a

The height of the image seen by the camera, in view space coordinates.

viewHeight should be a positive, non-zero value.

Accesses the Y component of viewArea

Leaves viewWidth unchanged

Adjusts viewDiagonal and viewAspect to match the new width.

viewAspect :: Floating a => Lens' (c a) a

The aspect ratio of the image seen by the camera. The ratio is W/H, i.e. values > 1.0 are wide and short, values < 1.0 are tall and narrow.

viewAspect should be a positive, non-zero value.

Leaves viewDiagonal unchanged

Adjusts viewArea, viewWidth, and viewHeight to have the new aspect ratio with the same diagonal size

viewDiagonal :: Floating a => Lens' (c a) a

The diagonal size of the image seen by the camera, in view space coordinates.

viewDiagonal should be a positive, non-zero value.

Leaves viewAspect unchanged

Scales viewArea, viewWidth, and viewHeight to match the new diagonal size with the same aspect ratio

class Camera c => Camera3D c where

The Camera3D class is an interface to cameras that exist in a 3D scene.

Note: The result of setting camera properties to values that do not satisfy the requirements specified for each field is undefined behavior. In particular, this may result in throwing exceptions and/or violating the lens laws.

Minimal complete definition

coordinateSystem, viewMatrix, invViewMatrix, projMatrix, invProjMatrix, rangeLimit, position, orientation

Methods

coordinateSystem :: Num a => Lens' (c a) (M33 a)

The base coordinate system for an unrotated camera.

rightward :: Num a => Lens' (c a) (V3 a)

The rightward axis, projected as the viewport positive X direction

This is almost always (V3 1 0 0).

upward :: Num a => Lens' (c a) (V3 a)

The upward axis, projected as the viewport positive Y direction

This is conventionally (V3 0 1 0) in 3D graphics programming, mapping the positive Y axis to the positive Y axis.

This is conventionally (V3 0 0 1) in many 3D modeling applications (e.g., Blender), so that the Z axis is vertical and the XY plane is horizontal.

forward :: Num a => Lens' (c a) (V3 a)

The forward axis, projected as viewport positive depth

This is conventionally the Z axis in 3D graphics programming, either (V3 0 0 1) for a left-handed coordinate system or (V3 0 0 (-1)) for right-handed. * This is conventionally (V3 0 1 0) in many 3D modeling applications (e.g., Blender), so that the Z axis is vertical and the XY plane is horizontal.

cameraMatrix :: (Epsilon a, RealFloat a) => Getter (c a) (M44 a)

Extract a transformation matrix from world space to view space.

For 3D cameras: cameraMatrix c ≡ projMatrix c !*! viewMatrix c

invCameraMatrix :: (Epsilon a, RealFloat a) => Getter (c a) (M44 a)

Extract a transformation matrix from view space to world space.

For 3D cameras: invCameraMatrix c ≡ invViewMatrix c !*! invProjMatrix c

viewMatrix :: (Epsilon a, RealFloat a) => Getter (c a) (M44 a)

Extract a transformation matrix from world space to the camera's local object space.

invViewMatrix :: (Epsilon a, RealFloat a) => Getter (c a) (M44 a)

Extract a transformation matrix from the camera's local object space to world space.

projMatrix :: (Epsilon a, RealFloat a) => Getter (c a) (M44 a)

Extract a transformation matrix from the camera's local object space to view space.

invProjMatrix :: (Epsilon a, RealFloat a) => Getter (c a) (M44 a)

Extract a transformation matrix from view space to the camera's local object space.

rangeLimit :: Lens' (c a) (a, a)

The depth range seen by the camera, in view space coordinates, as (near, far). Anything outside this range after projection to view space will be ignored.

Note: When using a depth buffer for rendering, the possible depth values will generally be distributed over this range, so avoid setting the range much larger than you need on pain of depth test bugs.

Both components should be non-zero, and nearLimit should be less than farLimit.

nearLimit is the near clipping plane.

farLimit is the far clipping plane.

nearLimit :: Lens' (c a) a

The near clipping plane, in view space coordinates.

Accesses the first component of rangeLimit

farLimit :: Lens' (c a) a

The far clipping plane, in view space coordinates.

Accesses the second component of rangeLimit

position :: Lens' (c a) (V3 a)

The position of the camera, in world space coordinates.

orientation :: Lens' (c a) (Quaternion a)

The world space orientation of the camera, as a unit quaternion.

orientation should have a magnitude of 1.

class Camera3D c => OrthographicCamera c where

The OrthographicCamera class is an interface to orthographic projection cameras.

Note: The result of setting camera properties to values that do not satisfy the requirements specified for each field is undefined behavior. In particular, this may result in throwing exceptions and/or violating the lens laws.

Minimal complete definition

orthoCamera, orthoFromPlanes, horizontalRange, verticalRange

Methods

orthoCamera

Arguments

:: RealFloat a

=> V2 a
Viewport size

-> a
Near clipping plane

-> a
Far clipping plane

-> c a

Create an orthographic camera with a view size and depth range. World space coordinates are assumed to match the viewport coordinates.

orthoFromPlanes

Arguments

:: RealFloat a

=> V2 a
Viewport size

-> a
Left

-> a
Right

-> a
Bottom

-> a
Top

-> a
Near

-> a
Far

-> c a

Create an orthographic camera from a view size and six clipping planes

horizontalRange :: Lens' (c a) (a, a)

The horizontal clipping planes.

Both components should be non-zero, and leftClip should be less than rightClip.

leftClip is the left clipping plane.

rightClip is the right clipping plane.

leftClip :: Lens' (c a) a

The left clipping plane, in view space coordinates.

Accesses the first component of horizontalRange

rightClip :: Lens' (c a) a

The right clipping plane, in view space coordinates.

Accesses the second component of horizontalRange

verticalRange :: Lens' (c a) (a, a)

The vertical clipping planes.

Both components should be non-zero, and bottomClip should be less than topClip.

bottomClip is the bottom clipping plane.

topClip is the top clipping plane.

bottomClip :: Lens' (c a) a

The bottom clipping plane, in view space coordinates.

Accesses the first component of verticalRange

topClip :: Lens' (c a) a

The top clipping plane, in view space coordinates.

Accesses the second component of verticalRange

class Camera3D c => PerspectiveCamera c where

The PerspectiveCamera class is an interface to perspective projection cameras. Which seems kind of redundant, and would be in real life, but some properties don't really make sense on (non-physical) orthographic cameras.

Note: The result of setting camera properties to values that do not satisfy the requirements specified for each field is undefined behavior. In particular, this may result in throwing exceptions and/or violating the lens laws.

Minimal complete definition

perspectiveCamera, focalLength, focalArea

Methods

perspectiveCamera

Arguments

:: RealFloat a

=> V2 a
Viewport size

-> CoordinateSystem a
Coordinate system

-> a
Focal length

-> a
Near clipping plane

-> a
Far clipping plane

-> c a

Create a perspective view camera.

focalLength :: Floating a => Lens' (c a) a

The "focal length" of the virtual camera.

focalLength should be a positive, non-zero value.

focalArea :: Lens' (c a) (Maybe (V2 a))

The "sensor" size of the virtual camera, relating FOV to focal length.

The default behavior will use viewArea with an additional scaling factor based on typical desktop monitor sizes &c. such that focalLength is roughly the 35mm equivalent focal length for a full-screen viewport. This means that:

In a full-screen window, focalLength will mean approximately the same thing that "focal length" does on many physical cameras.

The apparent size of rendered objects will not change if the viewport is resized.

If this is set to an explicit value, viewArea will be ignored and the specified size will be used (treated as millimeters). This means that:

Setting focalArea to Just (V2 36 24) will make focalLength behave as the 35mm equivalent value. (Standard 35mm film, of course, having an image size of 36mm x 24mm. What else would it be?)

The rendered scene will not change if the viewport is resized, and will be distorted if the aspect ratios don't match.

fieldOfView :: Floating a => Lens' (c a) (V2 (Angle a))

The angular field of view of the virtual camera. For "sensor" size d and focal length f, field of view is given by fov = 2 arctan(d / 2f).

Both dimensions should be positive, non-zero angles.

fovHorizontal and fovVertical are the X and Y components

Adjusts focalArea to match the new FOV

Leaves focalLength unchanged

fovHorizontal :: Floating a => Lens' (c a) (Angle a)

The angular field of view of the virtual camera along the horizontal view axis.

Accesses the X component of fieldOfView

Adjusts focalArea to match the new FOV

Leaves focalLength unchanged

fovVertical :: Floating a => Lens' (c a) (Angle a)

The angular field of view of the virtual camera along the vertical view axis.

Accesses the Y component of fieldOfView

Adjusts focalArea to match the new FOV

Leaves focalLength unchanged

effectiveFocalArea :: (PerspectiveCamera c, Floating a) => c a -> V2 a

Compute the effective focal area for a camera, either an explicitly set value or the default.

Camera modifier classes

class Camera3D c => JibCamera c where

Jib operations apply an offset to a camera in its local coordinate space.

Minimal complete definition

displacement

Methods

displacement :: Lens' (c a) (V3 a)

Displacement of camera relative to its base position.

distance :: Num a => Lens' (c a) a

Displacement along the camera's forward axis.

class Camera3D c => GimbalCamera c where

Gimbal operations apply a sequence of rotations to a camera in its local coordinate space.

Minimal complete definition

heading, elevation, roll

Methods

heading :: Lens' (c a) (Angle a)

Rotation angle in the horizontal plane

elevation :: Lens' (c a) (Angle a)

Rotation around the camera's horizontal axis

roll :: Lens' (c a) (Angle a)

Rotation around the camera's view direction

gimbalRotation :: (Epsilon a, RealFloat a) => Getter (c a) (Quaternion a)

All three rotations combined

invGimbalRotation :: (Epsilon a, RealFloat a) => Getter (c a) (Quaternion a)

The inverse of gimbalRotation.

Camera types

data OrthoCam a

Orthographic projection camera

data Cam a

Perspective projection camera

data Jib c a

Modifier that applies an offset to a camera in its local coordinate space.

data Gimbal c a

Modifier that applies a sequence of rotations to a camera in its local coordinate space.

Angles

data Angle a

Wrapper type representing an angle.

angleFrom :: AnIso' (Angle a) a -> a -> Angle a

Construct an angle using an angle unit isomorphism, e.g.:

angleFrom degrees 60 ≡ angleFrom turns (1/6) ≡ angleFrom radians (pi/3)

turns :: Iso' (Angle a) a

Converts between abstract Angles and "turns", or fractions of a complete circle.

radians :: Floating a => Iso' (Angle a) a

Converts between abstract Angles and radians.

degrees :: Fractional a => Iso' (Angle a) a

Converts between abstract Angles and degrees.

module Linear