Graphics.Camera.Internal

Documentation

axisVector :: (Foldable v, Additive v, Additive t, Num a) => v a -> Lens' (v (t a)) (t a)

toFocalLength :: Floating a => a -> Angle a -> a

Compute the focal length for a given sensor size and FOV angle.

toSensorSize :: Floating a => a -> Angle a -> a

Compute the sensor size for a focal length and FOV angle.

toFOV :: Floating a => a -> a -> Angle a

Compute the field of view for a given sensor size and focal length.

isoFocalLengthFOV :: Floating a => a -> Iso' a (Angle a)

For a constant sensor size we have an isomorphism between FOV and focal length

isoSensorFOV :: Floating a => a -> Iso' a (Angle a)

For a constant focal length we have an isomorphism between FOV and sensor size

module Graphics.Camera.Types