Graphics.Camera.Types

Documentation

data BaseCamera a

Generic 3D camera

Constructors

BCam

Fields

_bcamCoords :: M33 a

_bcamViewport :: V2 a

_bcamViewRange :: (a, a)

_bcamPosition :: V3 a

_bcamOrientation :: Quaternion a

bcamViewport :: forall a. Lens' (BaseCamera a) (V2 a)

bcamViewRange :: forall a. Lens' (BaseCamera a) (a, a)

bcamPosition :: forall a. Lens' (BaseCamera a) (V3 a)

bcamOrientation :: forall a. Lens' (BaseCamera a) (Quaternion a)

bcamCoords :: forall a. Lens' (BaseCamera a) (M33 a)

data OrthoCam a

Orthographic projection camera

Constructors

OCam

Fields

_ocamBaseCamera :: BaseCamera a

_ocamHRange :: (a, a)

_ocamVRange :: (a, a)

ocamVRange :: forall a. Lens' (OrthoCam a) (a, a)

ocamHRange :: forall a. Lens' (OrthoCam a) (a, a)

ocamBaseCamera :: forall a. Lens' (OrthoCam a) (BaseCamera a)

data Cam a

Perspective projection camera

Constructors

PCam

Fields

_pcamBaseCamera :: BaseCamera a

_pcamFocalArea :: Maybe (V2 a)

_pcamFocalLength :: a

pcamFocalLength :: forall a. Lens' (Cam a) a

pcamFocalArea :: forall a. Lens' (Cam a) (Maybe (V2 a))

pcamBaseCamera :: forall a. Lens' (Cam a) (BaseCamera a)

data Jib c a

Modifier that applies an offset to a camera in its local coordinate space.

Constructors

Jib

Fields

_jibCamera :: c a

_jibDisplacement :: !(V3 a)

jibDisplacement :: forall c a. Lens' (Jib c a) (V3 a)

jibCamera :: forall c a c. Lens (Jib c a) (Jib c a) (c a) (c a)

newJib :: Num a => c a -> Jib c a

data Gimbal c a

Modifier that applies a sequence of rotations to a camera in its local coordinate space.

Constructors

Gimbal

Fields

_gimbalCamera :: c a

_gimbalHeading :: !(Angle a)

_gimbalElevation :: !(Angle a)

_gimbalRoll :: !(Angle a)

gimbalRoll :: forall c a. Lens' (Gimbal c a) (Angle a)

gimbalHeading :: forall c a. Lens' (Gimbal c a) (Angle a)

gimbalElevation :: forall c a. Lens' (Gimbal c a) (Angle a)

gimbalCamera :: forall c a c. Lens (Gimbal c a) (Gimbal c a) (c a) (c a)

newGimbal :: Num a => c a -> Gimbal c a

data CoordinateSystem a

Describes the coordinate system of an unrotated camera.

Constructors

VerticalZ
Coordinate system common in 3D modeling applications. The positive Z axis points up and the positive Y axis points forward.

ZForward
Coordinate system common in 3D graphics programming. The positive Y axis points up and the positive Z axis points forward. This is a left-handed coordinate system.

ZBackward
Coordinate system common in 3D graphics programming. The positive Y axis points up and the positive Z axis points backward. This is a right-handed coordinate system.

OtherCoords (M33 a)
Arbitrary coordinate system, specified as (V3 rightward upward forward)